backend/requirements.txt
Django
djangorestframework
djangorestframework_simplejwt
drf-spectacular
django-allauth[socialaccount]==0.57.0
django-cors-headers
dj-rest-auth==7.0.0
python-dotenv
razorpay
django-jet-reboot
